refactor(field): rename field to schema
