--- src/field/index.js
@@ -1,2 +1,0 @@
-import field from './field';
-export default field;
