+++ src/schema/index.js
@@ -1,0 +1,2 @@
+import schema from './schema';
+export default schema;
